Fixing status code: 4901
I know that this is on the part of the Discord servers, but we must update the protocol, during the tests a week ago, this error in this protocol was significantly reduced. For some reason, the library Eris (Fork) is being forced to use old protocol which is not very good for the client side.